# remote-externalities: batch insert key/values