As suggested by @arkpar (https://github.com/paritytech/substrate/issues/13562#issuecomment-1520853103), adds batch inserting of key/value pairs to remote-externalities so the trie root doesn't need to be re-computed after every insertion.

## some (totally unscientific) try-runtime benchmarks

at `ws://localhost:9944` is an (optimised build) kusama node

| Task | Build | Before | After | Improvement |
|---------|-------|--------|-------|-----|
| `on-runtime-upgrade` (live) | release | 72s | 59s | 18% |
| `on-runtime-upgrade` (live) | debug | 174s | 73s | 58% |
| `create-snapshot` | release | 70s | 58s | 17% |
| `create-snapshot` | debug | 173s | 57s | 67% |
| `on-runtime-upgrade` (snap) | release | 21s | 5s | 76% |
| `on-runtime-upgrade` (snap) | debug | 125s | 19s | 84% |